feat: Enable IntegrationTestConfiguration in IntegrationTest annotation
This commit enables the `IntegrationTestConfiguration` class in the `IntegrationTest` annotation in the `IntegrationTest.java` file. The `IntegrationTestConfiguration` class was previously commented out, but now it is included in the `IntegrationTest` annotation, allowing it to be used during integration tests. This change ensures that the necessary configuration is applied for the integration tests to run successfully.